+++ src/lightwin/core/instructions_factory.py
@@ -141,7 +141,7 @@ class InstructionsFactory:
         if self._load_field_maps:
             field_maps = [elt for elt in elts if isinstance(elt, FieldMap)]
             self._field_factory.run_all(field_maps)
-            load_electromagnetic_fields(field_maps, True)
+            load_electromagnetic_fields(field_maps, cython=True)
 
         return instructions